+++ src/vs/editor/common/editorCommon.ts
@@ -2082,24 +2082,3 @@ export var Handler = {
 	Undo: 'undo',
 	Redo: 'redo',
 };
-
-/**
- *@internal
- */
-export function getSelectionSearchString(editor: ICommonCodeEditor): string {
-	let selection = editor.getSelection();
-
-	// if selection spans multiple lines, default search string to empty
-	if (selection.startLineNumber === selection.endLineNumber) {
-		if (selection.isEmpty()) {
-			let wordAtPosition = editor.getModel().getWordAtPosition(selection.getStartPosition());
-			if (wordAtPosition) {
-				return wordAtPosition.word;
-			}
-		} else {
-			return editor.getModel().getValueInRange(selection);
-		}
-	}
-
-	return null;
-}
