+++ src/vs/workbench/parts/search/browser/search.contribution.ts
@@ -24,7 +24,7 @@ import { IInstantiationService } from 'vs/platform/instantiation/common/instanti
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { IQuickOpenService } from 'vs/platform/quickOpen/common/quickOpen';
 import { ICodeEditorService } from 'vs/editor/common/services/codeEditorService';
-import { getSelectionSearchString } from 'vs/editor/common/editorCommon';
+import { getSelectionSearchString } from 'vs/editor/contrib/find/common/find';
 import { IViewletService } from 'vs/workbench/services/viewlet/browser/viewlet';
 import { KeyMod, KeyCode } from 'vs/base/common/keyCodes';
 import { ITree } from 'vs/base/parts/tree/browser/tree';
